[Thread Tutorial] Add Openthread starter code
