+++ examples/tutorials/thread_network/02_openthread/Makefile
